Don't inject failures inside libc calls
No need to inject failures when inside libc calls so ensure libc_force
is set around calls to the super function.

This reduces the number of iterations required to fully test all
branches (seeing many duplicates are removed) and will help the
dogfood test to behave a bit more predictably.